fix(appkit): extract kebabToCamel to a leaf module so the bundled CLI can reuse it
The earlier dedup pointed promote.ts at kebabToCamel in plugin.ts. That
resolved in-repo but broke the published tarball: the CLI is copied to
appkit's dist/cli flattened, and its `../../../../plugin` import pointed at a
dist/plugin.js that dist-appkit.ts never copies — so `npm run typegen` in the
template failed with ERR_MODULE_NOT_FOUND (caught by the PR Template Artifact
check).

Move kebabToCamel + KebabToCamel into a dependency-free leaf module
shared/src/naming.ts. plugin.ts re-exports them (so `from "shared"` consumers
like appkit.ts are unchanged), promote.ts imports from ../../../../naming, and
dist-appkit.ts copies naming.js next to the CLI in the tarball — mirroring how
it already copies schemas/. Verified by building the prerelease tarball and
loading the bundled promote.js: it now resolves naming.js and imports cleanly.

Genuine single-source (no duplicated helper) that survives bundling, rather
than dragging the heavy plugin.ts graph into the CLI bundle.
